refactor(adapter-openclaw): extract shared leaf utils into runtime-utils.ts
First cut of the runtime.ts god-file split (audit-2026-06 WS5). Moves the
20 pure, dependency-light JSON/object/string primitives shared across the
adapter's capability clusters (cron, sessions, channels, approvals,
activity summarization) into a new lib leaf module: stripAnsi/ANSI_RE,
firstString/firstStringAtPaths/stringAtPath, getJsonPath, isPlainObject/
isRecord/readPath/deepMerge, cloneJson, parseJsonValue/parseJsonLines,
readJsonFile, truncate, redactSensitiveText, slug, sleep, metadataValue/
metadataFiles.

This is the foundation the subsequent cluster extractions import, so they
never have to import back through runtime.ts (which would create a cycle).
Pure relocation — function bodies are byte-for-byte unchanged. None of
these were in the public/test export surface, so they're just imported
back; the class + the test seams (mergeChatStreams, the __*ForTest session
exports, SESSION_STORE_CACHE_MAX) stay exported from runtime.ts unchanged.

The cross-package deepMerge dedup with core/settings.ts (the plan's
"dedupe deepMerge" note) is behavior-touching across the core/adapter
boundary — deliberately NOT taken here; left as a follow-up.

runtime.ts 3,188 -> 3,053. Verified: typecheck, eslint (changed files),
adapter suite 115-0 (--isolate), full suite 5124-0, binary build (stamps
reverted), and a boot smoke against an openclaw PATH stub (adapter loads,
schedule activates via runtime.cron, GET /api/plugins/schedule/ -> 200).
